requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.2
